No unnecessary strings on POSIX header creation
The POSIX header creation methods created a new 512-byte string of NULL
bytes just to see if the header was an empty header. This could be simplified
with a single 512-byte constant string for comparison, but a short-circuiting
comparison will be more efficient over-all. Use the #each_byte enumerator with
the #any? test method on #nonzero? (data.each_byte.any?(&:nonzero?)).